move controller to its own file, clean and finish up styles for now
