+++ app/components/common/contributionsCategoryTable/contributionsCategoryTable.html
@@ -1,4 +1,4 @@
-<article class="odca-contentBlock--singleItem">
+<article class="odca-contributionsTable odca-contentBlock--singleItem">
 	<h3 class="odca-contributionsTable-categoryHeading">{{vm.contributionsCategory}} <span class="odca-contributionsTable-categoryHeading--categoryTotal">{{vm.totalContributions | currency:'$':0}}</span></h3>
 
 		<!--<pre>{{vm.contributions}}</pre>-->
